Rework session-file handling in loadOrCreateChatCompletionRequest
Refactor loadOrCreateChatCompletionRequest function to allow the use of non-existent sessionFile. Prior logic required the sessionFile to exist, but we now create a new instance if the provided file doesn't exist or if no sessionFile is given. This update improves flexibility and user experience of the ChatFlags functionality by providing more options for session continuity and preservation. And makes scripting with prior known file name easier.